requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
scipy
supabase
openai
pydantic
python-dotenv
python-multipart
slowapi
